src: fix `-Wformat` issues in trace calls
Also fixup `int` to `unsigned int` and adjust printf format strings.
These did not cause compiler warnings.

Cast large types to `long` to avoid dealing with printf masks for `size_t`
and other C99 types. Existing code often used `int` for this. I'll update
them to `long` in an upcoming commit.

+++ src/hostkey.c
@@ -106,7 +106,7 @@ hostkey_method_ssh_rsa_init(LIBSSH2_SESSION * session,
 #endif
     {
         _libssh2_debug((session, LIBSSH2_TRACE_ERROR,
-                       "unexpected rsa type: %.*s", type_len, type));
+                       "unexpected rsa type: %.*s", (int)type_len, type));
         return -1;
     }
